Don't remove watchers on permission change.
This can be far too slow (especially with membership inheritance) and notifications are not sent to watchers that are not allowed to view the item. If we still want to remove watchers that are no longer able to view the watched items, the redmine:watchers:prune task can be called periodically.

git-svn-id: http://redmine.rubyforge.org/svn/trunk@11300 e93f8b46-1217-0410-a6f0-8f06a7374b81